validaciones de cdeula y anios

+++ src/main/java/com/test/cognoware/bryam/parra/testBackend/serviceImpl/PersonaServiceImpl.java
@@ -6,9 +6,11 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.test.cognoware.bryam.parra.testBackend.model.Empresa;
 import com.test.cognoware.bryam.parra.testBackend.model.Persona;
 import com.test.cognoware.bryam.parra.testBackend.repository.PersonaRepository;
 import com.test.cognoware.bryam.parra.testBackend.service.PersonaService;
+import com.test.cognoware.bryam.parra.testBackend.validations.validations;
 
 
 @Service
@@ -19,6 +21,13 @@ public class PersonaServiceImpl implements PersonaService{
 
     @Override
     public Object createPersona(Persona persona) {
+        validations validation = new validations();
+        if( !validation.validaFecha(persona.getFecha_nacimiento())){
+            return "Fecha invalida";
+        }
+        if (!validation.validaCedula(persona.getCedula())) {
+            return "Cedula invalida";
+        }
         try {
             personaRepository.save(persona);
             return persona;
@@ -38,6 +47,13 @@ public class PersonaServiceImpl implements PersonaService{
 
     @Override
     public Object updatePersona(Persona persona) {
+        validations validation = new validations();
+        if( !validation.validaFecha(persona.getFecha_nacimiento())){
+            return "Fecha invalida";
+        }
+        if (!validation.validaCedula(persona.getCedula())) {
+            return "Cedula invalida";
+        }
         try {
             personaRepository.save(persona);
             return persona;
@@ -49,9 +65,18 @@ public class PersonaServiceImpl implements PersonaService{
     @Override
     public boolean deletePersona(Long id) {
         try {
-            personaRepository.deleteById(id);
+            Persona persona = this.personaRepository.findById(id).get();
+            if (persona != null) {
+                for (Empresa empresa : persona.getEmpresas()) {
+                    empresa.getPersonas().remove(persona);
+                }
+                persona.getEmpresas().clear();
+                personaRepository.deleteById(id);
+
+            }
             return true;
         } catch (Exception e) {
+            System.out.println("error eliminacion => "+e.getMessage());
             return false;
         }
     }
